Mid 9 - Added eventDispatcher to let others extend the framework.
After framework assambled the response, just before returning it,
the framework let the eventdispatcher class emit a ResponseEvent.

This response event contains the request and response objects.

Listeners to the event can access those objects and therefore can
modify the response (adding stuff to it).

Listeners need to be added to the eventdispatcher. A listener just
need to be of type callable (a function will do).

We can set the priority of the listeners to control which order
they will be alerted.